# docs/swagger.yaml
basePath: /
definitions:
  linkpulse_internal_models.CreateLinkRequest:
    properties:
      custom_alias:
        maxLength: 50
        minLength: 3
        type: string
      expires_at:
        type: string
      original_url:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - original_url
    type: object
  linkpulse_internal_models.ErrorDetails:
    properties:
      code:
        example: INVALID_CREDENTIALS
        type: string
      message:
        example: Invalid email or password
        type: string
    type: object
  linkpulse_internal_models.ErrorResponse:
    properties:
      detail:
        example: The email or password provided is incorrect.
        type: string
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/auth/login
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 401
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Invalid Credentials
        type: string
      type:
        example: https://linkpulse.com/errors/invalid-credentials
        type: string
    type: object
...
  linkpulse_internal_models.PaginatedResponse:
    properties:
      data: {}
      message:
        example: Items retrieved successfully
        type: string
      metadata:
        $ref: '#/definitions/linkpulse_internal_models.PaginationMetadata'
      request_id:
        example: uuid-request-id
        type: string
      success:
        example: true
        type: boolean
    type: object
  linkpulse_internal_models.PaginationMetadata:
    properties:
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  linkpulse_internal_models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      request_id:
        example: uuid-request-id
        type: string
      success:
        example: true
        type: boolean
    type: object
  linkpulse_internal_models.UpdateLinkRequest:
    properties:
      expires_at:
        type: string
      is_active:
        type: boolean
      title:
        maxLength: 255
        type: string
    type: object
...
  linkpulse_internal_models.ValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: The email field is required
        type: string
      rule:
        example: required
        type: string
    type: object
  linkpulse_internal_models.ValidationErrorResponse:
    properties:
      detail:
        example: Validation failed
        type: string
      details:
        items:
          $ref: '#/definitions/linkpulse_internal_models.ValidationError'
        type: array
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/links
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 422
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Validation Error
        type: string
      type:
        example: https://linkpulse.com/errors/validation-error
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: JyotiBhardwajj
    url: https://github.com/JyotiBhardwajj/LinkPulse
  description: |-
    Production-grade URL shortener with analytics, authentication, and observability.

    ## Authentication
    All protected endpoints require a Bearer JWT in the Authorization header.
    Obtain tokens via POST /api/v1/auth/login.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LinkPulse API
  version: "1.0"
paths:
...
  /api/v1/links:
    get:
      description: Returns a paginated list of all shortened links created by the
        authenticated user.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Search by URL or title
        in: query
        name: search
        type: string
      - description: 'Sort field: created_at, updated_at, expires_at'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc, desc'
        in: query
        name: order
        type: string
      - description: 'Filter by status: active, expired, inactive, deleted'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.PaginatedResponse'
        "304":
          description: Not Modified
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List links
      tags:
      - links
    post:
      consumes:
      - application/json
      description: Shortens a long URL and returns the short link details.
      parameters:
      - description: Link creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      security:
      - BearerAuth: []
      summary: Create short link
      tags:
      - links
  /api/v1/links/{id}:
    delete:
      description: Soft-deletes a shortened link. The short code will no longer resolve.
      parameters:
      - description: Link UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete link
      tags:
      - links
    get:
      description: Returns details of a specific shortened link owned by the authenticated
        user.
      parameters:
      - description: Link UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "304":
          description: Not Modified
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get link by ID
      tags:
      - links
    patch:
      consumes:
      - application/json
      description: Updates the title, expiry date, or active status of an existing
        link.
      parameters:
      - description: Link UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.UpdateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      security:
      - BearerAuth: []
      summary: Update link
      tags:
      - links
...
  /api/v1/links/{id}/stats:
    get:
      description: Returns total click count and basic metrics for a specific shortened
        link.
      parameters:
      - description: Link UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get link statistics
      tags:
      - links
  /api/v1/users/me:
    get:
      description: Returns the profile information of the currently authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /r/{code}:
    get:
      description: Redirects to the original URL for the given short code. Click analytics
        are recorded asynchronously.
      parameters:
      - description: Short code
        in: path
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Found
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      summary: Resolve short link
      tags:
      - redirect
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"